app/fraud/requirements.txt
requests==2.31.0
numpy==1.26.4
boto3==1.34.93
botocore==1.34.93
install==1.3.5
jmespath==1.0.1
psycopg2==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
pandas==2.2.2
scikit-learn==1.4.2
transformers==4.40.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
pgvector==0.2.5
kubernetes==29.0.0
